added FASTA file support to the Sequence IO module
[git-p4: depot-paths = "//sw/gpgpu/nvbio/main/": change = 18517243]

+++ nvbio/fasta/fasta_inl.h
@@ -124,7 +124,7 @@ uint8 FASTA_inc_reader::get()
 {
     if (m_buffer_pos >= m_buffer_size)
     {
-        m_buffer_size = uint32( gzread( m_file, &m_buffer[0], m_buffer.size() ) );
+        m_buffer_size = uint32( gzread( m_file, &m_buffer[0], (unsigned int)m_buffer.size() ) );
         m_buffer_pos  = 0;
     }
     return (m_buffer_pos < m_buffer_size) ? m_buffer[ m_buffer_pos++ ] : 255u;
@@ -230,7 +230,7 @@ uint8 FASTA_reader::get()
 {
     if (m_buffer_pos >= m_buffer_size)
     {
-        m_buffer_size = uint32( gzread( m_file, &m_buffer[0], m_buffer.size() ) );
+        m_buffer_size = uint32( gzread( m_file, &m_buffer[0], (unsigned int)m_buffer.size() ) );
         m_buffer_pos  = 0;
     }
     return (m_buffer_pos < m_buffer_size) ? m_buffer[ m_buffer_pos++ ] : 255u;
